requirements.txt
Flask==0.9
Flask-Script==0.3.3
Flask-Login==0.1.3
Flask-WTF==0.8

Flask-SQLAlchemy==0.16
redis==2.6.2
#hiredis==??? for up to 10x speed increase on parsing responses!

Flask-Uploads==0.1.3
boto==2.6.0

webassets==0.8
Flask-Assets==0.8
cssmin==0.1.4

# may be handled by Flask's dependencies as of 0.10
itsdangerous==0.17
